requirements.txt
torch
torchvision
torchaudio
transformers
streamlit
numpy
pandas
Pillow
scikit-learn
